+++ mps/test/function/66.c
@@ -90,7 +90,8 @@ static void test(void) {
   "create awl pool");
 
  cdie(
-  mps_pool_create(&poolmv, arena, mps_class_mv(), 0x4000, 128, 0x4000),
+  mps_pool_create(&poolmv, arena, mps_class_mv(),
+                  (size_t)0x4000, (size_t)128, (size_t)0x4000),
   "create mv pool");
 
  cdie(
